Port MXNet to CUDA 13 / cuDNN 9 / sm_120 (Blackwell)
This commit makes MXNet master build and run on NVIDIA Blackwell GPUs
(RTX PRO 4000 Blackwell, sm_120) with the CUDA 13.0 / cuDNN 9.14 /
NCCL 2.28 toolchain, driver R580, Python 3.11+, NumPy 1.26.

== Build system ==
* cmake/upstream/select_compute_arch.cmake: extend known arches for
  CUDA 11.8 (Ada, Hopper), 12.8 (Blackwell sm_100/sm_120), 13.0 (drop
  Maxwell/Pascal/Volta as CUDA 13 does); fix multi-digit-arch regex.
* CMakeLists.txt: bump CMAKE_CUDA_STANDARD 14 -> 17 (CCCL requires it);
  add ${CUDA_INC}/cccl to include path (CUDA 13 relocated Thrust/CUB
  under cccl/).

== CUDA 13 / cuDNN 9 API drift ==
* src/common/cuda/utils.h:
  - Relax CUDNN_VERSION_AS_STRING static_assert (cuDNN 9 allows
    CUDNN_MINOR >= 10).
  - Add CublasComputeTypeFromDataType helper for the cuBLAS 11+
    cublasComputeType_t signature change.
* src/imperative/cuda_graphs.h: switch to CUDA 12+ signature of
  cudaGraphExecUpdate (single cudaGraphExecUpdateResultInfo* out).
* src/operator/{linalg_impl.h, contrib/transformer.cu,
  quantization/quantized_fully_connected.cu}: wrap cublasGemmEx /
  cublasGemmStridedBatchedEx compute-type argument with
  CublasComputeTypeFromDataType under CUDA_VERSION >= 11000.
* src/operator/rnn-{inl.h,.cc,.cu}: gate all cuDNN RNN code paths
  behind CUDNN_VERSION < 9000.  cuDNN 9 removed the legacy RNN API
  (cudnnSetRNNDescriptor_v6, cudnnRNNForwardInference[Ex],
  cudnnRNNBackwardData[Ex], ...).  GPU RNN/LSTM/GRU is therefore
  unavailable in this build; see PERFORMANCE_NOTES.md.
* src/operator/numpy/np_unique_op.cu: drop deprecated
  thrust::unary_function base class (removed in Thrust 2.x).

== Real Blackwell / cuDNN 9 bugs found and fixed ==
* src/operator/tensor/indexing_op.h: HistogramEven upper_level was
  K+1 instead of K, producing bins of width (K+1)/K. Older CUB
  versions appeared to round integer samples to the nearest bin;
  CCCL 3.x (CUDA 13) does the float binning faithfully, which
  exposed the bug -- the gradient of take in 'clip' mode was being
  scattered to the wrong axis position.  Fixed both call sites.
  This made all 37 parameterised test_take[*] tests pass.
* src/operator/numpy/np_init_op.h: numpy_linspace_fwd cast int dtype
  by truncating toward zero (default C++ cast); NumPy's linspace
  floors toward -infinity for integer dtypes.  Switched to floor()
  before cast for integral DType.  Fixed all 160 linspace tests.
* src/operator/tensor/elemwise_binary_op_basic.cu: add explicit
  template instantiation of ElemwiseBinaryOp::DnsCsrDnsOp<plus/minus>
  for mshadow::Stream<gpu>*.  Under nvcc's separate-compilation
  pipeline the symbol otherwise wasn't emitted, and dependent .cc
  TUs ended up with undefined references.

== Namespace fallout (CCCL 3.x ::cuda) ==
CUDA 13 introduces a top-level ::cuda namespace via libcu++, which
collides with mxnet/mshadow's own 'cuda' sub-namespaces.  To avoid
the ambiguity:
* 3rdparty/mshadow/...: rename namespace mshadow::cuda -> cuda_impl
  (vendored copy, safe to rename).  Update all internal references.
* src/operator/<various>.cu: rename file-local 'namespace cuda { }'
  to 'namespace cuda_impl { }' inside namespace mshadow blocks.
* src/operator/{tensor/pseudo2DTranspose_op-inl.cuh,
  tensor/sort_op-inl.cuh, nn/depthwise_convolution_tf.cuh}: rename
  file-local 'namespace cuda' inside mxnet::op to 'cuda_kernels'.
* Mass-prefix bare 'cuda::X' references with 'mshadow::cuda_impl::X'
  in operator/cc/.cu files where applicable.

== Python library bugs ==
* python/mxnet/_deferred_compute.py: explicit ctypes argtypes for the
  three deferred-compute C entry points.  Python 3.11+ ctypes no
  longer auto-decays c_void_p*N arrays to void**, so the callees were
  reading uninitialised stack memory.  Fixed 3 deferred-compute tests
  that were failing with garbage Symbol.outputs.size() values.
* python/mxnet/numpy/multiarray.py: __array_ufunc__ fallback read
  inputs[0].device unconditionally; that fails when the first input
  is a plain numpy.ndarray.  Pick the device from any mxnet input.
* python/mxnet/numpy/fallback.py: filter out NumPy 1.20+-retired
  symbols (mirr/npv/pmt/... per NEP 32, alltrue/sometrue/msort,
  etc.).  The symbols still exist on import but raise RuntimeError
  when called.
* python/mxnet/test_utils.py: pass sum_check=False to scipy.stats
  .chisquare (scipy 1.10+ added a strict equality check on the
  observed/expected frequency totals that mxnet's bucketed
  generators routinely fail by ~1e-6).

== Wheel / packaging ==
* python/setup.py: install libmxnet.so as package_data inside the
  mxnet package directory rather than as a data_files entry.  The
  old layout placed the .so under <sysprefix>/mxnet/ which
  find_lib_path() doesn't search, so pip-installed wheels couldn't
  load the library.
* python/mxnet/libinfo.py: version bumped to 2.0.0+cu13.bw.<date>.

== Test compatibility ==
Across tests/python/{unittest,gpu,dnnl}: mechanical NumPy 1.24+
compat in tests:
* np.bool/np.int/np.float -> np.bool_/np.int_/np.float64 (deprecated
  aliases removed in NumPy 1.20).
* arr[list_of_slices] -> arr[tuple(list_of_slices)] (NumPy 1.24+
  rejects list-as-multi-index).
* np.repeat(list_of_Symbols, N) -> Python list comprehension (NumPy
  1.24+ refuses to build object arrays from heterogeneous input).
* Ragged nested-list literal in gen_broadcast_data -> plain Python
  list.
* type(out) == type(expected_out) comparison relaxed to also accept
  "both are tuples" (NumPy 1.25+ linalg.eig returns EigResult
  named-tuple, not plain tuple).
* tests/python/unittest/test_dgl_graph.py: arr.asnumpy()[0] ->
  arr.asnumpy().item() (NumPy now rejects [0] on 0-D arrays).
* tests/python/unittest/test_gluon_data.py: y.asscalar() ->
  y.item() (asscalar was deliberately removed in mxnet 2.0's numpy
  ndarray, the test was using a removed API).
* tests/python/unittest/test_gluon_probability_v2.py: cast inputs
  to scipy.stats.t to float64; scipy 1.10+ refuses to safe-cast
  float64 results back to float32.

== Test infrastructure ==
* New PERFORMANCE_NOTES.md describing where this build sits perf-wise
  (~2.7 TFLOP/s fp16 vs ~150 TFLOP/s tensor-core peak) and what would
  be needed for real Blackwell performance: cuBLASLt/CUTLASS paths,
  bf16 native support, cuDNN 9 RNN port, oneDNN bump, retuned launch
  configs.

== Test results ==
All ~13,000 GPU unit tests in tests/python/gpu/test_operator_gpu.py
pass on Blackwell after these fixes, modulo a single 5%-relative
numerical edge in test_pooling_versions (1 element of 480k, cuDNN 9
maxpool vs CPU reference - pre-existing tolerance issue, not a port
bug).  Six other test files have small numbers of pre-existing
upstream MXNet failures (custom op signature drift, sparse + np_shape
save/load, DataLoader collation under newer NumPy) which are
documented in PERFORMANCE_NOTES.md.

Build: cmake -DUSE_CUDA=ON -DUSE_CUDNN=ON -DUSE_OPENCV=ON
       -DUSE_ONEDNN=OFF -DMXNET_CUDA_ARCH=12.0 ..
       ninja mxnet
Toolchain: CUDA 13.0 + cuDNN 9.14 + NCCL 2.28 + driver 580.126.18.

+++ tests/python/unittest/test_gluon_probability_v2.py
@@ -996,7 +996,10 @@ def test_gluon_student_t():
             if hybridize:
                 net.hybridize()
             mx_out = net(df, loc, scale).asnumpy()
-            ss_f = ss.t(loc=0, scale=scale.asnumpy(), df=df.asnumpy())
+            # scipy >=1.10 refuses to safe-cast float64 results back into
+            # float32 underlying arrays; pass float64 inputs to match.
+            ss_f = ss.t(loc=0, scale=scale.asnumpy().astype('float64'),
+                        df=df.asnumpy().astype('float64'))
             if func == 'mean':
                 np_out = ss_f.mean()
             elif func == 'variance':
